Removed custom warnings and moved the IconWithLabel class to it's own file
