+++ Pod/Classes/Segments/IconWithLabel.swift
@@ -1,0 +1,18 @@
+//
+//  IconWithLabel.swift
+//  BetterSegmentedControl
+//
+//  Created by Arman Zoghi on 2/15/21.
+//
+
+import Foundation
+import UIKit
+
+public class IconWithLabel {
+    let icon: UIImage
+    let title: String
+    public init(icon: UIImage, title: String) {
+        self.icon = icon
+        self.title = title
+    }
+}
