+++ Pod/Classes/Segments/LabelSegment.swift
@@ -74,7 +74,6 @@ open class LabelSegment: BetterSegmentedControlSegment {
                     textColor: selectedTextColor,
                     accessibilityIdentifier: accessibilityIdentifier)
     }()
-    #warning("here")
     open func createLabel(withText text: String?,
                           backgroundColor: UIColor,
                           font: UIFont,
@@ -91,7 +90,6 @@ open class LabelSegment: BetterSegmentedControlSegment {
         label.accessibilityIdentifier = accessibilityIdentifier
         return label
     }
-    #warning("here")
 }
 
 public extension LabelSegment {
